Bump version to 1.1.0; add release notes and optimize load time with lazy loading for Monaco editor
