+++ RELEASE.md
@@ -1,0 +1,8 @@
+# Release notes
+# v1.1.0 (2024-12-8)
+- load time optimization
+18s -> 10s
+- lazy loading monaco editor
+
+## v1.0.0 (2024-11-26)
+- Initial release
