+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oligami/rustc-browser-wasi_shim",
-  "version": "0.1.0",
+  "version": "1.1.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
