+++ page/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rubrc",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "rubrc example page",
   "private": true,
   "author": "oligami <nziq53@gmail.com> (https://github.com/oligamiq)",
