+++ page/src/worker.ts
@@ -11,6 +11,8 @@ let ctx: Ctx;
 import RustcWorker from "./rustc?worker";
 let shared: SharedObject;
 
+const wasi_refs = [];
+
 globalThis.addEventListener("message", async (event) => {
   if (event.data.ctx) {
     rustc_worker = new RustcWorker();
@@ -38,9 +40,40 @@ globalThis.addEventListener("message", async (event) => {
         terminal(`loaded sysroot ${triple}\r\n`);
       })();
     }, ctx.load_additional_sysroot_id);
+
+    wasi_refs.push(wasi_ref);
+    if (wasi_refs.length === 2) {
+      setup_util_worker(wasi_refs, ctx);
+    }
   } else if (event.data.wasi_ref) {
     const { wasi_ref } = event.data;
 
     rustc_worker.postMessage({ wasi_ref_ui: wasi_ref });
+    wasi_refs.push(wasi_ref);
+    if (wasi_refs.length === 2) {
+      setup_util_worker(wasi_refs, ctx);
+    }
   }
 });
+
+import util_cmd_worker from "./util_cmd.ts?worker";
+import run_llvm_worker from "./llvm.ts?worker";
+
+const setup_util_worker = (
+  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  wasi_refs: any[],
+  ctx: Ctx,
+) => {
+  const util_worker = new util_cmd_worker();
+  const llvm_worker = new run_llvm_worker();
+
+  util_worker.postMessage({
+    wasi_refs,
+    ctx,
+  });
+
+  llvm_worker.postMessage({
+    wasi_refs,
+    ctx,
+  });
+}
